requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.2

# Trading & Finance
ccxt==4.2.25
yfinance==0.2.33
ta-lib==0.4.28
alpaca-py==0.13.3
binance-connector==3.5.1

# Web Scraping & Automation
selenium==4.16.0
beautifulsoup4==4.12.2
requests==2.31.0
playwright==1.40.0
scrapy==2.11.0

# AI & ML
openai==1.6.1
transformers==4.36.2
langchain==0.1.0
anthropic==0.8.1
google-generativeai==0.3.2

# Data Processing
sqlalchemy==2.0.23
redis==5.0.1
celery==5.3.4
pymongo==4.6.1

# Web Framework & API
fastapi==0.108.0
uvicorn==0.25.0
streamlit==1.29.0
plotly==5.18.0
dash==2.14.2

# Monitoring & Notifications
twilio==8.10.3
discord.py==2.3.2
python-telegram-bot==20.7
sendgrid==6.11.0

# Testing & Development
pytest==7.4.3
black==23.12.1
pylint==3.0.3